backend/requirements.txt
# FastAPI et dépendances web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Cache
redis==5.0.1
hiredis==2.2.3

# Manipulation de données
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.6.0
shap==0.44.1
tensorflow==2.15.0
joblib==1.3.2

# Indicateurs techniques
ta-lib==0.4.28
yfinance==0.2.28

# Corrélations et statistiques
statsmodels==0.14.0
seaborn==0.13.0

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilitaires
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
loguru==0.7.2

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Développement
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
